bit more refactor cleanup, need to finish validation next then document everything
