+++ src/index.ts
@@ -2,31 +2,8 @@
  * This is the entrypoint for SQL management
  */
 
-import { QueryParser, type ParseSQL } from "./query/parser/query.js"
-import type { CheckQuery } from "./query/validation/query.js"
-import type { SQLDatabaseSchema } from "./schema/database.js"
-
+export { getDatabase, type SQLDatabase } from "./database.js"
 export { createQueryBuilder } from "./query/builder/query.js"
 export type { ParseSQL } from "./query/parser/query.js"
 export { createDatabaseSchema } from "./schema/builder/database.js"
-
-/**
- * Starting point for a SQL Database
- */
-export interface SQLDatabase<Schema extends SQLDatabaseSchema> {
-  readonly schema: Schema
-
-  parseSQL<T extends string>(query: CheckQuery<T>): ParseSQL<T>
-}
-
-export function getDatabase<Schema extends SQLDatabaseSchema>(
-  schema: Schema
-): SQLDatabase<Schema> {
-  const parseSQL = <T extends string>(query: CheckQuery<T>): ParseSQL<T> => {
-    return new QueryParser(schema).parse(query as string) as ParseSQL<T>
-  }
-  return {
-    schema,
-    parseSQL,
-  }
-}
+export { SQLBuiltinTypes } from "./types.js"
